cleanup, factor out test module
