+++ modules/build.gradle.kts
@@ -13,7 +13,6 @@ subprojects {
         val implementation by configurations
         val annotationProcessor by configurations
         val testImplementation by configurations
-        val testRuntimeOnly by configurations
 
         // Auto service (SPI)
         annotationProcessor("com.google.auto.service:auto-service:1.0.1")
@@ -23,16 +22,7 @@ subprojects {
         implementation("com.github.Minestom:Minestom:d596992c0eafd8c")
 
         // Testing
-        testImplementation("org.junit.jupiter:junit-jupiter-api:5.8.1")
-        testImplementation("org.junit.jupiter:junit-jupiter-params:5.8.1")
-        testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:5.8.1")
-
-        testImplementation("org.testcontainers:testcontainers:1.17.3") {
-            exclude(group = "junit", module = "junit")
-        }
-        testImplementation("org.testcontainers:junit-jupiter:1.17.3") {
-            exclude(group = "junit", module = "junit")
-        }
+        testImplementation(project(":modules:test"))
     }
 
     tasks.getByName<Test>("test") {
