+++ modules/common/src/main/java/unnamed/mmo/util/EventUtil.java
