+++ modules/common/src/main/java/unnamed/mmo/util/FutureUtil.java
