+++ modules/development/src/main/java/unnamed/mmo/server/dev/Main.java
@@ -18,7 +18,6 @@ import org.bson.UuidRepresentation;
 import unnamed.mmo.chat.ChatManager;
 import unnamed.mmo.chat.storage.ChatStorage;
 
-import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ForkJoinPool;
 import java.util.concurrent.TimeUnit;
