+++ modules/test/README.md
@@ -1,0 +1,4 @@
+## test
+A module for test utilities used by the others. Only ever included for tests.
+
+Would eventually contain something similar to Minestoms internal command framework I guess
