+++ settings.gradle.kts
@@ -2,5 +2,6 @@ rootProject.name = "JointMMO"
 
 include(":modules")
 include(":modules:common")
+include(":modules:test")
 include(":modules:chat")
 include(":modules:development")
